requirements.txt
pillow
pyautogui
pyscreenshot
adb-tool-py==0.1.2
# development
wheel
sphinx
-e ./